move TestIntegrationV2 into _test file
